Move mocked_files to ert/conftest.py
